+++ pom.xml
@@ -24,11 +24,11 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
 		<!-- C2PA tool download configuration (macOS universal ZIP) -->
-    	<c2pa.version>v0.9.12</c2pa.version>
-    	<c2pa.url>https://github.com/contentauth/c2patool/releases/download/${c2pa.version}/c2patool-${c2pa.version}-universal-apple-darwin.zip</c2pa.url>
+    	<c2pa.version>v0.23.4</c2pa.version>
+    	<c2pa.url>https://github.com/contentauth/c2pa-rs/releases/download/c2patool-{c2pa.version}/c2patool-{c2pa.version}-universal-apple-darwin.zip</c2pa.url>
    		<c2pa.unpack.dir>${project.basedir}/tools</c2pa.unpack.dir>
     	<c2pa.unpack.archive>${c2pa.unpack.dir}/c2patool.zip</c2pa.unpack.archive>
-    	<!-- Folder name created by the ZIP for v0.9.12 (contains the binary named "c2patool") -->
+    	<!-- Folder name created by the ZIP for v0.23.4 (contains the binary named "c2patool") -->
     	<c2pa.unzipped.folder>c2patool-${c2pa.version}-universal-apple-darwin</c2pa.unzipped.folder>
 	</properties>
 
@@ -228,8 +228,6 @@
 					</execution>
 				</executions>
 			</plugin>
-
-			
 		</plugins>
 	</build>
 </project>
