+++ src/test/java/dev/coms4156/project/metadetect/UserServiceTest.java
@@ -1,9 +1,10 @@
-package dev.coms4156.project.metadetect.service;
+package dev.coms4156.project.metadetect;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import dev.coms4156.project.metadetect.service.UserService;
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.Map;
